move getHostSystemName to util
